Fix SetDword build error: use unsigned long instead of DWORD in header
The RegistryReader.h header does not include <Windows.h>, so DWORD is
not defined there. Use 'unsigned long' (which DWORD is a typedef for)
in both the header declaration and .cpp implementation.

Also add missing dwDisposition parameter to RegCreateKeyEx call
(requires 9 arguments, not 8).

+++ CppClient/CppClient/RegistryReader.h
@@ -48,5 +48,5 @@ public:
 	bool SetWString(const std::wstring& name, const std::wstring& value) noexcept;
 
 	// Write a DWORD value to the registry
-	bool SetDword(const std::wstring& name, DWORD value) noexcept;
+	bool SetDword(const std::wstring& name, unsigned long value) noexcept;
 };
